refactor(runtime): extract bus worker to orchestrator/worker module
Move run_worker, run_worker_filtered, submit_turn, register_token_sink,
and register_extensions into a dedicated worker.rs sub-module. Convert
orchestrator.rs to a directory module (orchestrator/mod.rs) so worker
methods stay as impl Orchestrator in the sub-file.

mod.rs shrinks by ~280 lines. All 45 tests pass.
